requirements.txt
# Requirements for dataset creation

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0

# HuggingFace transformers for tokenizer
transformers>=4.30.0
torch>=2.0.0  # Required by transformers

# For HuggingFace Inference API (caption generation)
requests>=2.31.0

# Optional: for better performance
# accelerate>=0.20.0

